refactoring - retiring scripts/utils.js
