+++ .eslintignore
@@ -1,4 +1,4 @@
 node_modules/*
-deps/*
-scripts/fallback.js
+events/deps/*
+events/scripts/fallback.js
 web-test-runner.config.js
